fix: email, authkey가 아닌 id로 삭제하도록 변경...

+++ src/main/java/com/luckkids/domain/confirmEmail/ConfirmEmailRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ConfirmEmailRepository extends JpaRepository<ConfirmEmail, Long> {
+public interface ConfirmEmailRepository extends JpaRepository<ConfirmEmail, Integer> {
     Optional<ConfirmEmail> findByEmail(String email);
     Optional<ConfirmEmail> findByEmailAndAuthKey(String email, String authkey);
 }
